preprocessingws/app/requirements.txt
Flask
numpy
textract
nltk
geograpy3
pdftotext
langdetect
